IsoBHPlot(IsoGene)	R Documentation

Plot of adjusted p-values using BH or BY adjustment

Description

The function produces a plot with adjusted p-values using BH (Benjamini and Hochberg 1995) and BY (Benjamini and Yekutilie 2004) procedures controlling for FDR. The raw p-values and adjusted BH and BY p-values are plotted.

Usage


IsoBHPlot(rp, FDR, stat = c("E2", "Williams", "Marcus",
"M", "ModifM"))

Arguments

rp 	 raw p-value matrix with each row for one gene and 6 columns, the first column contains the Probe.ID, the second to the sixth columns are raw p-values for the five test statistics

FDR 	 the desired FDR to control

stat 	 choose one of the five test statistic to use

Value

A plot of adjusted p-values using BH and BY procedures will be produced.

Author(s)

Lin et al.

References

Testing for Trend in Dose-Response Microarray Experiments: a Comparison of Testing Procedures, Multiplicity, and Resampling-Based Inference, Lin et al. 2007, Stat. App. in Gen. & Mol. Bio., 6(1), article 26.

See Also

IsoTestBH, IsoRawp

Examples


  rp <- data.frame(paste("g", 1:100), matrix(runif(500,0,1), 100, 5))
  IsoBHPlot(rp, FDR = 0.05, stat = "E2")

[Package IsoGene version 1.0-15 Index]