IsoGene-package(IsoGene)	R Documentation

IsoGene

Description

Library IsoGene aims to identify for genes with a monotonic trend in the expression levels with respect to the increasing doses using several test statistics. They include the global likelihood ratio test (E^2, Bartholomew 1961, Barlow et al. 1972 and Robertson et al. 1988), Williams (1971, 1972), Marcus (1976), the M (Hu et al. 2005) and the modified M (Lin et al. 2007). The p-values of the five test statistics are obtained using permutation and they are adjusted using BH (Benjamini and Hochberg 1995) and BY (Benjamini and Yekutilie 2004) procedures are used for controlling the FDR.

Details

Package: 	 IsoGene

Type: 	 Package

Version: 	 1.0

Date: 	 2007-05-02

License: 	 Free

Value

The package includes the following functions:

IsoGene1 	 calculates the five test statistics in testing both increasing and decreasing alternatives for a single gene

IsoGenem 	 calculates the five test statistics in testing both increasing and decreasing alternatives for all the genes in the data set

IsoRawp 	 obtains the raw (one-sided and two-sided) p-values using permuations

IsoTestBH 	 BH or BY procedure to adjust p-values while controlling FDR

IsoGenemSAM 	 calculates the SAM test statistic

Isofudge 	 calculates the fudge factor in the SAM test statistic

Isoqqstat 	 calculates the SAM test statistic using permutations

Isoallfdr 	 obtains the delta table in the SAM procedure

Isoqval 	 the SAM procedure to obtain q-values

IsoTestSAM 	 the SAM procedure to obtain a list of significant genes

IsoSAMPlot 	 SAM plot

IsoBHPlot 	 plot of adjusted BH and BY p-values

IsoPlot 	 plot of data, sample means, and a fitted isotonic regression curve with a likely direction

IsopvaluePlot 	 plot of p-values obtained using permutation under increasing or decreasing alternatives

Author(s)

Lin et al.

Maintainer: Dan Lin <dan.lin@uhasselt.be>

References

Testing for Trend in Dose-Response Microarray Experiments: a Comparison of Testing Procedures, Multiplicity, and Resampling-Based Inference, Lin et al. 2007, Stat. App. in Gen. & Mol. Bio., 6(1), article 26.

See Also

mt.rawp2adjp, IsoGene1, IsoGenem,IsoRawp, IsoTestBH, IsoGenemSAM, Isofudge, Isoqqstat,Isoallfdr,Isoqval,IsoTestSAM, IsoSAMPlot, IsoBHPlot, IsoPlot.

[Package IsoGene version 1.0-13 Index]