IsoGene1(IsoGene)	R Documentation

The five test statistics calculated for both the increasing and decreasing trends

Description

The function calculates the values for the five test statistics (the global likelihood test, Williams, Marcus, M, and the modified M) for testing increasing and decreasing alternatives.

Usage


IsoGene1(x, y)

Arguments

x 	 indicates the dose levels

y 	 is the gene expression for one gene

Value

A list with components

E2.up 	 the test statistic of the global likelihood test for testing increasing alternative.

Williams.up 	 the test statistic of Williams for testing increasing alternative.

Marcus.up 	 the test statistic of Marcus for testing increasing alternative.

M.up 	 the M test statistic for testing increasing alternative.

ModM.up 	 the test statistic of the modified M for testing increasing alternative.

E2.dn 	 the test statistic of Williams for testing decreasing alternative.

Williams.dn 	 the test statistic of global likelihood test for testing decreasing alternative.

Marcus.dn 	 the test statistic of Williams for testing decreasing alternative.

M.dn 	 the test statistic of global likelihood test for testing decreasing alternative.

ModM.dn 	 the test statistic of Williams for testing increasing alternative.

direction 	 the direction with the higher likelihood of increasing (indicated by "u") or decreasing (indicated by "d") trend.

Note

This function calculates the five test statistics for both increasing and decreasing ordered alternatives for a single gene.

Author(s)

Lin et al.

References

Testing for Trend in Dose-Response Microarray Experiments: a Comparison of Testing Procedures, Multiplicity, and Resampling-Based Inference, Lin et al. 2007, Stat. App. in Gen. & Mol. Bio., 6(1), article 26.

See Also

isoreg

Examples


  x <- c(rep(1,3),rep(2,3),rep(3,3),rep(4,3))
  y <- c(rnorm(3,1,1),rnorm(3,2,1),rnorm(3,3,1),rnorm(3,4,1))
  stat <- IsoGene1(x,y)
  stat

[Package IsoGene version 1.0-12 Index]