IsoGenemSAM(IsoGene)	R Documentation

The five SAM test statistics calculated for both the increasing and decreasing trends

Description

The function calculates the values for the five SAM test statistics (the global likelihood test, Williams, Marcus, M, and the modified M) for the most likely direction.

Usage


IsoGenemSAM(x, y, fudge.factor)

Arguments

x 	 indicates the dose levels

y 	 gene expression for all genes

fudge.factor 	 the fudge factor values to be used in the SAM test statistics

Value

A list with components

E2 	 the SAM test statistic of global likelihood test for the likely direction of each gene.

Williams 	 the test statistic of Williams for the likely direction of each gene.

Marcus 	 the test statistic of Marcus for the likely direction of each gene.

M 	 the M test statistic for the likely direction of each gene.

ModM 	 the test statistic of the modified M for the likely direction of each gene.

direction 	 the direction with the higher likelihood of increasing (indicated by "u") or decreasing (indicated by "d") trend.

Note

This function calculates the five test statistics for both increasing and decreasing ordered alternatives for all the genes (rows in the data set).

Author(s)

Lin et al.

References

Testing for Trend in Dose-Response Microarray Experiments: a Comparison of Testing Procedures, Multiplicity, and Resampling-Based Inference, Lin et al. 2007, Stat. App. in Gen. & Mol. Bio., 6(1), article 26.

See Also

isoreg, IsoGene1, Isofudge

[Package IsoGene version 1.0-12 Index]