IsoPlot(IsoGene)	R Documentation

IsoPlot

Description

Plot of the data points and the sample means at each dose

Usage


IsoPlot(x, y, type=c("continuous", "ordinal"), add.curve = FALSE)

Arguments

x 	 indicates the dose levels

y 	 is the gene expression for one gene

type 	 specifies the dose levels to "continuous" or "ordinal". The default is "continous".

add.curve 	 specifies whether a fitted isotonic regression curve with a likely direction is added or not. The default is FALSE.

Value

Plot of the data points, the sample means for each dose (either as continuous or ordinal), and a fitted isotonic regression curve (optional) is produced.

Note

This function produces a plot for a single gene.

Author(s)

Lin et al.

Examples


  x <- c(rep(1,3), rep(2,3), rep(3,3), rep(4,3))
  y <- c(rnorm(3,1,1), rnorm(3,2,1), rnorm(3,3,1), rnorm(3,4,1))
  IsoPlot(x, y)
  IsoPlot(x, y, type="ordinal", add.curve=TRUE)

[Package IsoGene version 1.0-12 Index]