IsoRawp(IsoGene)	R Documentation

IsoRawp

Description

The function calculates the raw one-sided and two-sided p-values for each test statistic using permutations.

Usage


IsoRawp(x, y, niter)

Arguments

x 	 numeric vector containing the dose levels

y 	 a data frame of the gene expression with Probe IDs as row names

niter 	 number of permutations to use

Details

The number of permutations to use can be chosen based on the number of possible permutations of samples. If the possible number is too big, usually >5000 permutations can be sufficient.

Value

A list of components

raw.p.one 	 returns the one-sided p-value matrix for the five test statisticsin in 6 columns: the first column is the probe ID, the second to the last columns contain the raw p-values for each test statistic

raw.p.two 	 returns the two-sided p-value matrix for the five test statistics in 6 columns: the first column is the probe ID, the second to the last columns contain the raw p-values for each test statistic

rawp.up 	 returns the one-sided p-value matrix testing increasing alternative for the five test statistics in 6 columns: the first column is the probe ID, the second to the last columns contain the raw p-values for each test statistic

rawp.dn 	 returns the one-sided p-value matrix testing decreasing alternative for the five test statistics in 6 columns: the first column is the probe ID, the second to the last columns contain the raw p-values for each test statistic

Note

For each gene, the one-sided p-values are calculated from min(p^Up, p^Down) and the two sided p-values are calculated from min{2 * min(p^Up, p^Down), 1}, where p^Up and p^Down are the p-values calculated for each ordered alternative.

Author(s)

Lin et al.

References

Lin et al. (2007). Microarray Experiments: a Comparis. on of Testing Procedures, Multiplicity, and Resampling-Based Inference, Stat. App. in Gen. & Mol. Bio., 6(1), article 26.

See Also

IsoTestBH

Examples


## Not run: 
  set.seed(1234)
  x <- c(rep(1,3),rep(2,3),rep(3,3))
  y1 <- matrix(rnorm(90, 1,1),10,9)  # 10 genes with no trends
  y2 <- matrix(c(rnorm(30, 1,1), rnorm(30,2,1),
                 rnorm(30,3,1)), 10, 9)  # 10 genes with increasing trends
  y <- data.frame(rbind(y1, y2))  # y needs to be a data frame 
  rp <- IsoRawp(x, y, niter = 1000)
  rp
## End(Not run)

[Package IsoGene version 1.0-15 Index]