IsoSAMPlot(IsoGene)	R Documentation

Plots produced using the SAM procedure

Description

The function produces four plots using the SAM procedure for one of the five test statistics (the likelihood ratio test, Williams, Marcus, the M and modified M tests): FDR vs. delta, number of significant genes vs. delta, number of false positives vs. delta, and the observed vs. expected SAM test statistics obtained from permutations.

Usage



IsoSAMPlot(qqstat, allfdr, FDR, stat)

Arguments

qqstat 	 output from function Isoqqstat containing the test statistics of permutations

allfdr 	 the delta table obtained from function Isoallfdr

FDR 	 choose the desired FDR to control

stat 	 choose one of the five test statistics to use

Value

returns four plots produced using the SAM procedure.

Note

This function produces four plots using the SAM procedure for the five test statistics. To use the SAM procedure, the number of genes in the dataset is preferably larger than 500.

Author(s)

Lin et al.

See Also

isoreg, Isoqqstat, Isoallfdr,Isoqval, IsoTestSAM

Examples


  set.seed(1234)
  x <- c(rep(1,3),rep(2,3),rep(3,3))
  y1 <- matrix(rnorm(9000, 1,1),1000,9) ## 1000 genes with no trends
  y2 <- matrix(c(rnorm(3000, 1,1),rnorm(3000,2,1),rnorm(3000,3,1)),1000,9) ## 1000 genes with increasing trends
  y <- data.frame(rbind(y1, y2)) ##y needs to be a data frame
  qqstat <- Isoqqstat(x, y, fudge="pooled", niter=100)
  allfdr <- Isoallfdr(qqstat, , stat = "E2")
  IsoSAMPlot(qqstat, allfdr, FDR = 0.1, stat = "E2")

[Package IsoGene version 1.0-12 Index]