IsoTestBH(IsoGene)	R Documentation

Test of monotonic trends using the five test statistics with BH or BY adjustment

Description

The function adjusts for the raw p-values of the five test statistics using BH or BY procedure.

Usage


IsoTestBH(rp, FDR, type = c("BH", "BY"), stat = c("E2",
"Williams", "Marcus", "M", "ModifM"))

Arguments

rp 	 raw p-value matrix with each row for one gene and 6 columns, the first column contains the Probe.ID, the second to the sixth columns are raw p-values for the five test statistics

FDR 	 the desired FDR to control

type 	 choose BH or BY procedure to control FDR

stat 	 choose one of the five test statistics to use

Details

The input raw p-values to this function can be the one sided or the two sided ones which are obtained using function raw.p. The results using one sided p-values and FDR controlling at alpha/2 is equivalent to that using two sided p-values and FDR controlling at alpha.

Value

sign.genes 	 A list of significant genes while controlling FDR is obtained, with 4 columns: the first column is the probe ID, the second column is the row id, the third column is the raw p-values of the significant genes and the last column is the adjusted p-values of significant genes using BH or BY procedure

Note

This function only allows one type of FDR adjustment, either BH or BY. For other type of adjustment, see function mt.rawp2adjp in package multtest.

Author(s)

Lin et al.

See Also

'mt.rawp2adjp', IsoRawp

Examples


  set.seed(1234)
  rp <- data.frame(paste("g",1:100), matrix(runif(500,0,0.1), 100, 5))
  sign <- IsoTestBH(rp, FDR = 0.05, type = "BH", stat = "E2")

[Package IsoGene version 1.0-12 Index]