IsoTestSAM(IsoGene)	R Documentation

Obtaining the list of significant genes using the SAM procedure

Description

The function obtains the list of significant genes using the SAM procedure for the five test statistics (the global likelihood test, Williams, Marcus, M, and the modified M).

Usage


IsoTestSAM(x, y, fudge, niter, FDR, stat)

Arguments

x 	 numeric vector containing the dose levels

y 	 data frame of the gene expression with Probe ID as row names

fudge 	 option used for calculating the fudge factor in the SAM test statistic, either "pooled" (fudge factor will be automatically computed in the function), or "none" if no fudge factor is used

niter 	 number of permutations to use

FDR 	 choose the desired FDR to control

stat 	 choose one of the five test statistics to use

Value

sign.genes1: a list of genes declared significant using the SAM procedure in a matrix of 5 columns. The first colomn is the probe id, the second column is the corresponding row number of the probe in the dataset, and the third column is the ordered test statistic values, and the fourth column is the q-values of the SAM procedure. The last two columns are raw p-values based on permutations and BH adjusted p-values.

Note

This function obtains the list of significant genes using the SAM procedure for the five test statistics. To use the SAM procedure, the number of genes in the dataset is preferably larger than 500.

Author(s)

Lin et al.

See Also

isoreg, Isofudge, IsoGenemSAM, Isoqqstat, Isoallfdr,Isoqval, IsoSAMPlot

Examples


  set.seed(1234)
  x <- c(rep(1,3),rep(2,3),rep(3,3))
  y1 <- matrix(rnorm(9000, 1,1),1000,9) ## 1000 genes with no trends
  y2 <- matrix(c(rnorm(3000, 1,1),rnorm(3000,2,1),rnorm(3000,3,1)),1000,9) ## 1000 genes with increasing trends
  y <- data.frame(rbind(y1, y2)) ##y needs to be a data frame
  SAM.obj <- IsoTestSAM(x, y, fudge="pooled", niter=100, FDR=0.05, stat="E2") 

[Package IsoGene version 1.0-12 Index]