Isoallfdr(IsoGene)	R Documentation

Obtaining the delta table in the SAM procedure

Description

The function obtains the delta table in the SAM procedure for the five test statistics (the global likelihood test, Williams, Marcus, M, and the modified M).

Usage


Isoallfdr(qqstat, ddelta, stat)

Arguments

qqstat 	 output from function Isoqqstat containing the test statistics of permutations

ddelta 	 give a list of values as cut-off to find the number of significant genes in the SAM procedure. If unspecified, the default value is assigned using the centiles of the absolute difference between the observed and expected test statistics.

stat 	 choose one of the five test statistics to use

Value

dtable: the delta table in the SAM procedure containing six columns. The first column is the cut-off value to find the number of significant genes, the second column is the meidan number of false positives, the third column is the 90% percentile number of false positives, the fourth column is the number of significant genes, the fifth column is the median FDR, and the last column is the 90% FDR.

Note

This function calculates the delta table in the SAM procedure for the five test statistics. To use the SAM procedure, the number of genes in the dataset is preferably larger than 500.

Author(s)

Lin et al.

See Also

isoreg, Isoqqstat, Isoqval,IsoTestSAM, IsoSAMPlot

Examples


  set.seed(1234)
  x <- c(rep(1,3),rep(2,3),rep(3,3))
  y1 <- matrix(rnorm(4500, 1,1),500,9) ## 500 genes with no trends
  y2 <- matrix(c(rnorm(1500, 1,1),rnorm(1500,2,1),rnorm(1500,3,1)),500,9) ## 500 genes with increasing trends
  y <- data.frame(rbind(y1, y2)) ##y needs to be a data frame
  qqstat <- Isoqqstat(x, y, fudge="pooled",niter=100)
  allfdr <- Isoallfdr(qqstat,,stat="E2")
 

[Package IsoGene version 1.0-14 Index]