IsopvaluePlot(IsoGene)	R Documentation

Plot of p-values from permutations under increasing or decreasing alternatives

Description

The function calculates the p-values using permutations under increasing and decreasing ordered alternatives for one gene. The p-values (p^up and p^down) are obtained from the plot of null distribution and observed statistics.

Usage


IsopvaluePlot(x, y, niter, stat = c("E2", "Williams", "Marcus", "M", "ModifM"))

Arguments

x 	 the dose levels

y 	 the gene expressions

niter 	 the number of permutations to use

stat 	 choose one of the five test statistics to use

Value

Plots of the null distribution and the observed test statistic under increasing and decreasing ordered alternatives.

Note

The function obtains the p-values under increasing and decreasing ordered alternatives for a single gene.

Author(s)

Lin et al.

See Also

IsoGene1

Examples


x <- c(rep(1,3), rep(2,3), rep(3,3), rep(4,3))
y <- c(rnorm(3,1,1), rnorm(3,2,1), rnorm(3,3,1), rnorm(3,4,1)) 

IsopvaluePlot(x, y, niter = 1000, stat = "Williams")

[Package IsoGene version 1.0-15 Index]