Isoqval(IsoGene)	R Documentation

Obtaining the list of significant genes using the SAM procedure

Description

The function obtains the list of significant genes using the SAM procedure for the five test statistics (the global likelihood test, Williams, Marcus, M, and the modified M).

Usage


Isoqval(delta, allfdr, qqstat, stat)

Arguments

delta 	 the delta value as cut-off to find the number of significant genes

allfdr 	 the delta table obtained from function Isoallfdr

qqstat 	 output from function Isoqqstat containing the test statistics of permutations

stat 	 choose one of the five test statistics to use

Value

A list of components

res 	 returns the list genes with descending q-values of the SAM procedure in three columns: the first column is the row number of the genes, the second column is the observed test statistic values, and the last column is the q-values

sign.list 	 returns the list of significant genes found by the defined delta value with descending p-values in three columns: the first column is the row number of the genes, the second column is the observed test statistic values, and the last column is the q-values

Note

This function obtains the list of significant genes using the SAM procedure for the five test statistics. To use the SAM procedure, the number of genes in the dataset is preferably larger than 500.

Author(s)

Lin et al.

See Also

isoreg, Isoqqstat, Isoallfdr, IsoTestSAM, IsoSAMPlot

Examples


  set.seed(1234)
  x <- c(rep(1,3),rep(2,3),rep(3,3))
  y1 <- matrix(rnorm(9000, 1,1),1000,9) ## 1000 genes with no trends
  y2 <- matrix(c(rnorm(3000, 1,1),rnorm(3000,2,1),rnorm(3000,3,1)),1000,9) ## 1000 genes with increasing trends
  y <- data.frame(rbind(y1, y2)) ##y needs to be a data frame
  qqstat <- Isoqqstat(x, y, fudge="pooled", niter=100)
  allfdr <- Isoallfdr(qqstat, ,stat="E2")
  qval <- Isoqval(delta=0.2, allfdr, qqstat, stat="E2")

[Package IsoGene version 1.0-12 Index]